+++ hydrometrics-main-back/src/main/java/com/service/hydrometrics/models/DTO/alert/ConfirmAlertDTO.java
@@ -2,5 +2,5 @@ package com.service.hydrometrics.models.DTO.alert;
 
 import java.io.Serializable;
 
-public record ConfirmAlertDTO(long alertId, String dataCamp, Double prediction) implements Serializable {
+public record ConfirmAlertDTO(long alertId, String dataCamp, Integer prediction) implements Serializable {
 }
