+++ hydrometrics-main-back/src/main/java/com/service/hydrometrics/models/DTO/graphics/DataDTO.java
@@ -2,6 +2,6 @@ package com.service.hydrometrics.models.DTO.graphics;
 
 import java.io.Serializable;
 
-public record DataDTO(String dateTime, Double value) implements Serializable {
+public record DataDTO(String dateTime, Integer value) implements Serializable {
 
 }
