+++ hydrometrics-main-back/src/main/java/com/service/hydrometrics/utils/WeatherDataMapper.java
@@ -1,11 +1,11 @@
 package com.service.hydrometrics.utils;
 
 public record WeatherDataMapper(String date_time,
-                                Double precipitation,
-                                Double relative_humidity,
-                                Double solar_radiation,
-                                Double temperature,
-                                Double wind_direction,
-                                Double wind_speed,
+                                Integer precipitation,
+                                Integer relative_humidity,
+                                Integer solar_radiation,
+                                Integer temperature,
+                                Integer wind_direction,
+                                Integer wind_speed,
                                 long station_id) {
 }
